+++ R/llplot.r
@@ -260,7 +260,7 @@ lty, lwd, col, level=99.99, refline=0, ...) {
          id0  <- id0[not.na]
          id00 <- id00[not.na]
          k    <- length(ai)
-         warning(mstyle$warning("Studies with NAs omitted from plotting."), call.=FALSE)
+         warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "studies", "study"), "with NAs omitted from plotting.")), call.=FALSE)
       }
 
       if (na.act == "na.fail")