+++ R/tes.r
@@ -224,7 +224,7 @@ tes <- function(x, vi, sei, subset, data,
             yi <- yi[not.na]
             vi <- vi[not.na]
             theta <- theta[not.na]
-            warning(mstyle$warning("Studies with NAs omitted from test."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "studies", "study"), "with NAs omitted from test.")), call.=FALSE)
 
          }