+++ R/to.long.r
@@ -660,7 +660,7 @@ data, slab, subset, add=1/2, to="none", drop00=FALSE, vlong=FALSE, append=TRUE, 
             slab <- slab[not.na]
             if (has.data)
                data <- data[not.na,]
-            warning(mstyle$warning("Tables with NAs omitted."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "tables", "table"), "with NAs omitted.")), call.=FALSE)
          }
 
          if (na.act == "na.fail")
@@ -750,7 +750,7 @@ data, slab, subset, add=1/2, to="none", drop00=FALSE, vlong=FALSE, append=TRUE, 
             slab <- slab[not.na]
             if (has.data)
                data <- data[not.na,]
-            warning(mstyle$warning("Tables with NAs omitted."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "tables", "table"), "with NAs omitted.")), call.=FALSE)
          }
 
          if (na.act == "na.fail")
@@ -840,7 +840,7 @@ data, slab, subset, add=1/2, to="none", drop00=FALSE, vlong=FALSE, append=TRUE, 
             slab <- slab[not.na]
             if (has.data)
                data <- data[not.na,]
-            warning(mstyle$warning("Tables with NAs omitted."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "tables", "table"), "with NAs omitted.")), call.=FALSE)
          }
 
          if (na.act == "na.fail")
@@ -930,7 +930,7 @@ data, slab, subset, add=1/2, to="none", drop00=FALSE, vlong=FALSE, append=TRUE, 
             slab <- slab[not.na]
             if (has.data)
                data <- data[not.na,]
-            warning(mstyle$warning("Tables with NAs omitted."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "tables", "table"), "with NAs omitted.")), call.=FALSE)
          }
 
          if (na.act == "na.fail")
@@ -991,7 +991,7 @@ data, slab, subset, add=1/2, to="none", drop00=FALSE, vlong=FALSE, append=TRUE, 
             slab <- slab[not.na]
             if (has.data)
                data <- data[not.na,]
-            warning(mstyle$warning("Tables with NAs omitted."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "tables", "table"), "with NAs omitted.")), call.=FALSE)
          }
 
          if (na.act == "na.fail")
@@ -1049,7 +1049,7 @@ data, slab, subset, add=1/2, to="none", drop00=FALSE, vlong=FALSE, append=TRUE, 
             slab <- slab[not.na]
             if (has.data)
                data <- data[not.na,]
-            warning(mstyle$warning("Tables with NAs omitted."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "tables", "table"), "with NAs omitted.")), call.=FALSE)
          }
 
          if (na.act == "na.fail")
@@ -1104,7 +1104,7 @@ data, slab, subset, add=1/2, to="none", drop00=FALSE, vlong=FALSE, append=TRUE, 
             slab <- slab[not.na]
             if (has.data)
                data <- data[not.na,]
-            warning(mstyle$warning("Tables with NAs omitted."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "tables", "table"), "with NAs omitted.")), call.=FALSE)
          }
 
          if (na.act == "na.fail")
@@ -1188,7 +1188,7 @@ data, slab, subset, add=1/2, to="none", drop00=FALSE, vlong=FALSE, append=TRUE, 
             slab <- slab[not.na]
             if (has.data)
                data <- data[not.na,]
-            warning(mstyle$warning("Tables with NAs omitted."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "tables", "table"), "with NAs omitted.")), call.=FALSE)
          }
 
          if (na.act == "na.fail")
@@ -1244,7 +1244,7 @@ data, slab, subset, add=1/2, to="none", drop00=FALSE, vlong=FALSE, append=TRUE, 
             slab <- slab[not.na]
             if (has.data)
                data <- data[not.na,]
-            warning(mstyle$warning("Tables with NAs omitted."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "tables", "table"), "with NAs omitted.")), call.=FALSE)
          }
 
          if (na.act == "na.fail")
@@ -1309,7 +1309,7 @@ data, slab, subset, add=1/2, to="none", drop00=FALSE, vlong=FALSE, append=TRUE, 
             slab <- slab[not.na]
             if (has.data)
                data <- data[not.na,]
-            warning(mstyle$warning("Tables with NAs omitted."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "tables", "table"), "with NAs omitted.")), call.=FALSE)
          }
 
          if (na.act == "na.fail")
@@ -1396,7 +1396,7 @@ data, slab, subset, add=1/2, to="none", drop00=FALSE, vlong=FALSE, append=TRUE, 
             slab <- slab[not.na]
             if (has.data)
                data <- data[not.na,]
-            warning(mstyle$warning("Tables with NAs omitted."), call.=FALSE)
+            warning(mstyle$warning(paste(sum(has.na), ifelse(sum(has.na) > 1, "tables", "table"), "with NAs omitted.")), call.=FALSE)
          }
 
          if (na.act == "na.fail")